Updated repo to current working version in prod
Fixed a bug where sometimes the names would get cut off in the ID3 tagging, leaving only the first letter.

Made MPDTube turn on random mode if the mode was previously disabled.

MPDtube (+Nurdbot) will now communicate back to the channel where the !mpdtube command was executed from. (This also works for private messages)

nfs_delay has been changed to nfs_timeout. Mpdtube will now check every second if the file exists, and continues as soon as it does. 
Otherwise, it will time out and throw an error. This should significantly improve the speed of MPDtube, as it no longer waits a set amount of time.